+++ frontend/src/index.css
@@ -146,7 +146,7 @@ body {
 
 .footer h3 {
   font-size: 1.2em;
-   margin-bottom: 1em;
+  margin-bottom: 1em;
 }
 
 .download-data {
@@ -350,6 +350,13 @@ button:disabled {
   cursor: pointer;
   border-radius: 5px;
 }
+
+#saveMarkers:disabled {
+  background-color: grey;
+  cursor: not-allowed;
+  pointer-events: none; /* Disable mouse events */
+}
+
 #disableMarkerAddingButton {
   background-color: #dc3545; /* Red color */
   padding: 10px;
