+++ frontend/src/services/axiosInstance.js
@@ -4,7 +4,7 @@ const API_URL = 'https://denniesbor.com';
 
 const axiosInstance = axios.create({
     baseURL: API_URL,
-    timeout: 5000,
+    timeout: 10000,
     headers: {
         'Content-Type': 'application/json',
         'Accept': 'application/json',
