+++ frontend/src/utils/persistMarkers.jsx
@@ -1,29 +1,25 @@
 import { saveMarkersAsGeoJSON } from "./saveFileUtils";
 import { getCurrentUser, getUserDetails } from "../services/authService";
 import axiosInstance from "../services/axiosInstance";
+import formatData from "./formatMarkers2GeoJSON";
 
 const VITE_API_URL = "https://denniesbor.com";
 
-const markerLabels = {
-  three_ph_transformer: "3 Ph. Transf",
-  circuit_breaker: "Circuit Breaker",
-  single_ph_transformer: "Single Ph. Transf",
-  primary_power_line: "Pri. Pwr Line",
-  sec_power_line: "Sec. Pwr Line",
-  control_building: "Control Room",
-};
-
 async function saveMarkers(payload) {
   try {
-    await axiosInstance.post(
+    const response = await axiosInstance.post(
       `${VITE_API_URL}/gis/bulk_update/update_markers/`,
       payload
     );
+    console.log("Markers saved successfully:", response.data);
+    return { success: true };
   } catch (error) {
     if (error.response) {
       console.error("Error saving markers:", error.response.data);
+      return { success: false, error: error.response.data };
     } else {
       console.error("Error saving markers:", error.message);
+      return { success: false, error: error.message };
     }
   }
 }
@@ -49,73 +45,13 @@ export async function saveMarkersToDatabase(
     return;
   }
 
-  const validMarkers = markers[selectedSubstation.SS_ID].filter((marker) => {
-    if (typeof marker.getLabel === "function") {
-      const label = marker.getLabel().text;
-      return typeof label === "string" && label.trim() !== "";
-    } else {
-      const label = marker.label.text;
+  const { geojson, markersToSave } = formatData(
+    markers,
+    selectedSubstation,
+    allowEmpty
+  );
 
-      return typeof label === "string" && label.trim() !== "";
-    }
-  });
-
-
-  if (
-    (!markers[selectedSubstation.SS_ID] ||
-      markers[selectedSubstation.SS_ID].length === 0 ||
-      validMarkers.length === 0) &&
-    !allowEmpty
-  ) {
-    console.error("Markers list is empty or contains invalid labels.");
-    return;
-  }
-
-  const markersToSave = allowEmpty
-    ? markers[selectedSubstation.SS_ID]
-    : validMarkers;
-
-  const geojson = {
-    type: "FeatureCollection",
-    features: [
-      {
-        type: "Feature",
-        geometry: {
-          type: "MultiPoint",
-          coordinates: markersToSave.map((marker) => [
-            marker.getPosition().lng(),
-            marker.getPosition().lat(),
-          ]),
-        },
-        properties: {
-          SS_ID: selectedSubstation.SS_ID,
-          SS_NAME: selectedSubstation.SS_NAME,
-          SS_OPERATOR: selectedSubstation.SS_OPERATOR,
-          SS_TYPE: selectedSubstation.SS_TYPE,
-          SS_VOLTAGE: selectedSubstation.SS_VOLTAGE,
-          connected_tl_id: selectedSubstation.connected_tl_id,
-          LINE_VOLTS: selectedSubstation.LINE_VOLTS,
-          REGION: selectedSubstation.REGION,
-          REGION_ID: selectedSubstation.REGION_ID,
-          threePhaseTransformerCount: 0,
-          singlePhaseTransformerCount: 0,
-          primaryPowerLineCount: 0,
-          secondaryPowerLineCount: 0,
-          totalTransformerCount: 0,
-          markers: markersToSave.map((marker) => {
-            const labelObj = marker.getLabel();
-            const labelKey = labelObj.text;
-            const attributes = marker.attributes;
-            return {
-              label: labelKey,
-              attributes: attributes,
-              color: labelObj.color,
-            };
-          }),
-        },
-      },
-    ],
-  };
+  console.log("GEojson data", geojson);
 
   if (downloadChecked) {
     saveMarkersAsGeoJSON(geojson, selectedSubstation);
@@ -138,6 +74,12 @@ export async function saveMarkersToDatabase(
         };
       }),
     };
-    await saveMarkers(payload);
+
+    const result = await saveMarkers(payload);
+    if (result.success) {
+      alert("Markers saved successfully.");
+    } else {
+      alert(`Error saving markers: ${result.error}`);
+    }
   }
 }
